fix bug

+++ init.sql
@@ -4,7 +4,7 @@ CREATE TABLE member
     id BIGINT NOT NULL PRIMARY KEY,
     name VARCHAR(50) NOT NULL,
     age INT NOT NULL,
-    CGPA DECIMAL NOT NULL,
+    CGPA FLOAT NOT NULL,
     CONSTRAINT CGPA_range CHECK(CGPA >= 0 AND CGPA <= 4.0),
     CONSTRAINT age_range CHECK(age >= 18 and age <= 22)
 );
